tabs work smoother, css animations
